ui: add watchlist to navbar and remove watchlist dynamic feature
