+++ .idea/gradle.xml
@@ -12,7 +12,6 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/core" />
-            <option value="$PROJECT_DIR$/watchlist" />
           </set>
         </option>
       </GradleProjectSettings>
