+++ .idea/misc.xml
@@ -30,12 +30,12 @@
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/font/mulish_medium.xml" value="0.20572916666666666" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/font/mulish_semibold.xml" value="0.20572916666666666" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/activity_detail.xml" value="0.17552083333333332" />
-        <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/activity_main.xml" value="0.22" />
+        <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/activity_main2.xml" value="0.14130434782608695" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/activity_splash.xml" value="0.1414324569356301" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/fragment_home.xml" value="0.1414324569356301" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/fragment_search.xml" value="0.16666666666666666" />
-        <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/fragment_watchlist.xml" value="0.1414324569356301" />
+        <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/fragment_watchlist.xml" value="0.1" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/item_movie.xml" value="0.22" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/item_movie_big.xml" value="0.3333333333333333" />
         <entry key="..\:/Users/galih/OneDrive/Documents/Documents/A. Software Engineering/Android Development/Projects/FilmKu/app/src/main/res/layout/item_movie_big_load.xml" value="0.22" />
