+++ app/src/main/java/com/giftech/filmku/presentation/main/home/HomeFragment.kt
@@ -47,7 +47,6 @@ class HomeFragment : Fragment() {
         if(activity != null){
             setupAdapter()
             getData()
-            setButtonClick()
         }
     }
