+++ app/src/main/java/com/giftech/filmku/presentation/main/watchlist/WatchlistViewModel.kt
@@ -1,16 +1,14 @@
-package com.giftech.filmku.watchlist
+package com.giftech.filmku.presentation.main.watchlist
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import com.giftech.filmku.core.domain.usecase.FilmUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
-
-class WatchlistViewModel (
-    useCase: FilmUseCase
+@HiltViewModel
+class WatchlistViewModel @Inject constructor(
+    private val useCase: FilmUseCase
 ): ViewModel() {
     val watchlist = useCase.getWatchlist().asLiveData()
 }
