+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -11,4 +11,9 @@
         android:icon="@drawable/search_selector"
         android:title="@string/title_search" />
 
+    <item
+        android:id="@+id/navigation_watchlist"
+        android:icon="@drawable/watchlist_selector"
+        android:title="@string/title_watchlist" />
+
 </menu>
