--- watchlist/build.gradle
@@ -1,31 +1,0 @@
-plugins {
-    id 'com.android.dynamic-feature'
-    id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
-    id 'com.google.dagger.hilt.android'
-}
-android {
-    compileSdk 32
-
-    defaultConfig {
-        minSdk 21
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles 'proguard-rules.pro'
-        }
-    }
-    buildFeatures{
-        viewBinding true
-    }
-}
-
-apply from: '../shared_dependencies.gradle'
-
-dependencies {
-    implementation project(":app")
-    implementation project(":core")
-}
