--- watchlist/src/main/AndroidManifest.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:dist="http://schemas.android.com/apk/distribution"
-    package="com.giftech.filmku.watchlist">
-
-    <dist:module
-        dist:instant="false"
-        dist:title="@string/title_watchlist">
-        <dist:delivery>
-            <dist:install-time />
-        </dist:delivery>
-
-        <dist:fusing dist:include="true" />
-    </dist:module>
-
-    <application>
-        <activity
-            android:name=".WatchlistActivity"
-            android:exported="true" >
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW"/>
-                <category android:name="android.intent.category.DEFAULT"/>
-                <data android:scheme="filmku"
-                    android:host="watchlist"/>
-            </intent-filter>
-        </activity>
-    </application>
-
-</manifest>
