--- watchlist/src/main/java/com/giftech/filmku/watchlist/WatchlistComponent.kt
@@ -1,20 +1,0 @@
-package com.giftech.filmku.watchlist
-
-import android.content.Context
-import com.giftech.filmku.di.WatchlistModuleDependencies
-import dagger.BindsInstance
-import dagger.Component
-
-@Component(dependencies = [WatchlistModuleDependencies::class])
-interface WatchlistComponent {
-
-    fun inject(activity: WatchlistActivity)
-
-    @Component.Builder
-    interface Builder {
-        fun context(@BindsInstance context: Context): Builder
-        fun appDependencies(watchlistModuleDependencies: WatchlistModuleDependencies): Builder
-        fun build(): WatchlistComponent
-    }
-
-}
