--- watchlist/src/main/res/layout/item_movie.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="horizontal"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <com.google.android.material.card.MaterialCardView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="@dimen/_8sdp"
-        android:layout_marginLeft="@dimen/_24sdp"
-        app:cardElevation="@dimen/_8sdp"
-        app:cardCornerRadius="@dimen/_6sdp">
-
-        <ImageView
-            android:id="@+id/iv_poster"
-            android:scaleType="centerCrop"
-            android:foreground="?android:attr/selectableItemBackgroundBorderless"
-            android:layout_width="@dimen/_85sdp"
-            android:layout_height="@dimen/_120sdp"
-            tools:srcCompat="@tools:sample/avatars" />
-    </com.google.android.material.card.MaterialCardView>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingLeft="@dimen/_16sdp"
-        android:paddingRight="@dimen/_24sdp"
-        android:orientation="vertical">
-        <TextView
-            android:id="@+id/tv_title"
-            style="@style/item_title_big"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/_8sdp"
-            android:text="Spiderman: No Way Home" />
-
-        <TextView
-            android:id="@+id/tv_vote"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginVertical="@dimen/_8sdp"
-            android:paddingHorizontal="@dimen/_4sdp"
-            android:layout_marginTop="@dimen/_4sdp"
-            android:drawableLeft="@drawable/ic_star_12"
-            android:drawablePadding="@dimen/_4sdp"
-            style="@style/item_vote_big"
-            android:text="9.1/10 IMDb"/>
-
-        <com.google.android.material.chip.ChipGroup
-            android:id="@+id/container_genre"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:chipSpacingVertical="@dimen/_8sdp">
-
-
-        </com.google.android.material.chip.ChipGroup>
-    </LinearLayout>
-
-</LinearLayout>
